Changed cascading to be more simplified, but working well now. See description for important NB.
Instead of choosing cascading settings for on start, on pause, on end session, you just decide cascading settings and the default behavior applies.

I changed ComponentActivityStatus to just be a boolean, active or inactive (true or false).
So pomodoro is active when it is in working stage and not "initial", otherwise it is inactive.
So music plays and blocker stays active while in working stage, even if paused, and automatically stops when in a break.

If I want to change that later, there is code right before this commit that is relevant.