Changed outline group to disclosure group AND fixed sidebar, now uses hashable + \.self instead of ID because the view update problem is built into swiftui.  AND some more things, see comment
Doesn't change unless identifiable changes, and identifiable = id
So no chagne id = no change view
Tried a few things but it was ugly
I was already using \.self on outlinegroup for this reason. But then I was using ID as the tag, therefore it was not actually changing the "selection", therefore I could not "go to [workspace]". That is now fixed.

Finished up statistics view. List view looking good now.